Create grafo funcional
You are playing a game consisting of n planets. Each planet has a teleporter to another planet (or the planet itself).

You have to process 
q
q queries of the form: You are now on planet 
a
a and want to reach planet 
b
b. What is the minimum number of teleportations?